grpcserver/requirements.txt
Django==3.1.4
djangogrpcframework==0.2
djangorestframework==3.12.2
grpcio==1.34.0
grpcio-tools==1.34.0
mysqlclient==2.0.2